[XPLO] any exploration method can save output from method, plus aligne Betab SA and stochanalysis with standard explo
