+++ gama.core/src/gama/core/kernel/batch/exploration/sobol/SobolExploration.java
@@ -164,14 +164,7 @@ public class SobolExploration extends AExplorationAlgorithm {
 		sobol_analysis.evaluate();
 
 		/* Save the simulation values in the provided .csv file (input and corresponding output) */
-		if (hasFacet(IKeyword.BATCH_OUTPUT)) {
-			String path_to = Cast.asString(scope, outputFilePath.value(scope));
-			final File f = new File(FileUtils.constructAbsoluteFilePath(scope, path_to, false));
-			final File parent = f.getParentFile();
-			if (!parent.exists()) { parent.mkdirs(); }
-			if (f.exists()) { f.delete(); }
-			sobol_analysis.saveSimulation(f);
-		}
+		if (hasFacet(IKeyword.BATCH_OUTPUT)) { saveRawResults(scope, res_outputs); }
 
 		/* Save the Sobol analysis report in a .txt file */
 		String path_to = Cast.asString(scope, getFacet(IKeyword.BATCH_REPORT).value(scope));
